RF: Split off core Dataset
